# client: Optimize tests