--- a/packages/client/test/integration/StreamrClient.test.ts
+++ b/packages/client/test/integration/StreamrClient.test.ts
@@ -44,7 +44,7 @@ describeRepeats('StreamrClient', () => {
     let streamDefinition: { id: string, partition: number }
 
     beforeAll(async () => {
-        const stream = await createPartitionedTestStream()
+        const stream = await createPartitionedTestStream(module)
         streamParts = createStreamPartIterator(stream)
     })