--- a/packages/protocol/src/utils/StreamPartID.ts
+++ b/packages/protocol/src/utils/StreamPartID.ts
@@ -1,7 +1,7 @@
 import { StreamID, toStreamID } from "./StreamID"
 
 const DELIMITER = '#'
-const MAX_PARTITION_COUNT = 100
+export const MAX_PARTITION_COUNT = 100
 
 export type StreamPartID = string & { readonly __brand: 'streamPartID' } // Nominal typing