--- a/packages/protocol/src/utils/index.ts
+++ b/packages/protocol/src/utils/index.ts
@@ -5,7 +5,7 @@ import { createTrackerRegistry, TrackerRegistry, SmartContractRecord } from "./T
 import { generateMnemonicFromAddress, parseAddressFromNodeId } from './NodeUtil'
 import { keyToArrayIndex } from "./HashUtil"
 import { StreamID, toStreamID, StreamIDUtils } from "./StreamID"
-import { StreamPartID, toStreamPartID, StreamPartIDUtils } from "./StreamPartID"
+import { StreamPartID, toStreamPartID, StreamPartIDUtils, MAX_PARTITION_COUNT } from "./StreamPartID"
 import { EthereumAddress, ENSName } from "./types"
 
 export {
@@ -25,6 +25,7 @@ export {
     StreamIDUtils,
     StreamPartID,
     StreamPartIDUtils,
+    MAX_PARTITION_COUNT,
     EthereumAddress,
     ENSName
 }