Optimized some client tests to use partitioned streams. That way we don't need need to create multiple streams in a single test. 

Speed-up:
- `Subscriber.test.ts`: 152s -> 39s
- `StreamrClient.test.ts`: 126s -> 26s (24s+2s)

In the updated `Subscriber.test.ts` and `StreamrClient.test.ts` we create one stream with 100 partitions. Each test case then "consumes" one partition and executes the test operations for that partition. Added some helper functions to make it easy to manage the test partitions (`createPartitionedTestStream((`, `createStreamPartIterator()` in `utils.ts`).

Subscribing and publishing don't need independent streams in these tests, because each subscribe/publish operation targets a single stream partition. Any operations to other partitions should not interfere the functionality. In that sense these modifications actually add some test coverage for subscriber/publisher implementations.

Extracted also some partitioning-related test cases from `StreamrClient.test.ts` to separate test (`Partitions.test.ts`). That tests uses the fake environment as partition computation/selection doesn't need real streams. 

### Future improvements

We can use this partitioning style also in some other tests. It is not applicable for all tests, since many functionalities have some stream-level aspects, e.g. permissions and storage node.